% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_lineage.R
\name{ancestor_matrix}
\alias{ancestor_matrix}
\title{Compute ancestor matrix}
\usage{
ancestor_matrix(g, mode = "out")
}
\arguments{
\item{g}{a graph}
}
\value{
a sparse logical matrix where each row list all ancestors of a node
}
\description{
Compute ancestor matrix
}
