% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aris.R
\docType{data}
\name{aris}
\alias{aris}
\title{aris example}
\format{
An object of class \code{"numeric"}
}
\source{
This dataset was generated during lineage analysis by LINEAGE.
}
\usage{
data("aris")
}
\description{
A dataframe containing ARI values among effective subspaces.
}
