% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{correct}
\alias{correct}
\title{correct}
\usage{
correct(x)
}
\arguments{
\item{x}{A dataframe containing mitochondrial variants frequency matrix where a column represented a single cell and a row represented variants frequency of a specific mitochondrial genotype. Required.}
}
\value{
A dataframe containing a modified mitochondrial genotype matrix.
}
\description{
Using mitochondrial variants frequency matrix as input, transform the zeroes into ones when the median of the non-zero frequencies in the same row >= 0.6.
}
\details{
Mitochondrial genotype matrix modification
}
\examples{
data("TF1_clones")
data=TF1_clones$data
d=data_prepare(data)
d2=data.frame()
for(i in 1:12){
    di=d[[i]]
    di2=as.data.frame(t(apply(di,1,correct)))
    d2=rbind(d2,di2)
}

}
