% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_lineage_session}
\alias{create_lineage_session}
\title{Create a new lineage session}
\usage{
create_lineage_session(namespace, config)
}
\arguments{
\item{namespace}{Namespace for organizing lineage}

\item{config}{Configuration returned by lineage_config}
}
\value{
A list containing the session information
}
\description{
Create a new lineage session
}
