% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs.R
\docType{data}
\name{cs}
\alias{cs}
\title{cs example}
\format{
An object of class \code{"numeric"}
}
\source{
This dataset was generated during lineage analysis by LINEAGE.
}
\usage{
data("cs")
}
\description{
Indexes of selected subspaces.
}
