% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_prepare}
\alias{data_prepare}
\title{data_prepare}
\usage{
data_prepare(data)
}
\arguments{
\item{data}{A dataframe containing mitochondrial variants frequency matrix and two other column "altAllele" and "refAllele". 
"altAllele" column represents the mutant allele; "refAllele" column represents the reference allele. Required.}
}
\value{
A list containing 12 submatrices with different mutation types. ("altAllele" column and "refAllele" column are removed)
}
\description{
Split the mitochondrial genotype matrix into 12 submatrices according to mutation types.
}
\details{
Matrix splitting based on mutation types
}
\examples{
data("TF1_clones")
data=TF1_clones$data
d=data_prepare(data)

}
