% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diss.R
\docType{data}
\name{diss}
\alias{diss}
\title{diss example}
\format{
An object of class \code{"list"}
}
\source{
This dataset was generated during lineage analysis by LINEAGE.
}
\usage{
data("diss")
}
\description{
A list containing Euclidean distance matrixes of selected subspaces.
}
