% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{finalize_lineage}
\alias{finalize_lineage}
\title{Finalize the lineage session}
\usage{
finalize_lineage(session)
}
\arguments{
\item{session}{The lineage session object}
}
\value{
The unique run ID for this lineage process
}
\description{
Finalize the lineage session
}
