% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_markers}
\alias{find_markers}
\title{find_markers}
\usage{
find_markers(f, cluster)
}
\arguments{
\item{f}{A list containing the frequency distribution of a variant. Required.}

\item{cluster}{A list containing the consensus clustering result. Required.}
}
\value{
A dataframe containing AUC score and p-value of the given variant.
}
\description{
find_markers performs evaluation of variant as marker for the consensus clustering result by calculating the receiver operating characteristics (AUC) score and Pearson's correlation coefficient between the binary cluster labels and the frequency distribution.
}
\details{
Marker variants identification
}
\examples{
data("TF1_clones")
data("scc")
data=TF1_clones$data
d=data_prepare(data)
tops2=data.frame()
for(i in 1:12){
    di=d[[i]]
    tops2=rbind(tops2,top_sites(di)[[2]])
}
find_markers_2=function(x,cluster=scc){
    a=find_markers(as.numeric(as.character(x)),cluster=as.numeric(as.character(scc)))
    a=a[a$p==min(a$p),]
    return(a)
}
markers=apply(tops2,1,find_markers_2)

}
