% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_lineage.R
\name{get_cells}
\alias{get_cells}
\title{Internal helper function to build a cells tibble from pseudotime and identity_score matrix.}
\usage{
get_cells(pseudotime, identity_scores)
}
\arguments{
\item{pseudotime}{a numeric vector of pseudotime value for each cell}

\item{identity_scores}{a numeric matrix of cell identity scores with number of
row matching length(pseudotime). Each row give identity scores of the cell
in each cluster. The higher the score the more likely the cell belong to the cluster.}
}
\value{
a tibble with combined informations
}
\description{
It is useful to check parameter values and produce uniformized output.
}
