% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{grid_2d}
\alias{grid_2d}
\title{Generate a uniform 2D grid around space defined by x,y points}
\usage{
grid_2d(x, y, nx = 30, ny = 30, filter = TRUE)
}
\arguments{
\item{x}{x coordinate of the points in 2D space}

\item{y}{y coordinate of the points in 2D space}

\item{nx}{number of grid point on x-axis}

\item{ny}{number of grid point on y-axis}

\item{filter}{a boolean, when TRUE remove grid point that are not the direct
nearest neighbor of a given point}
}
\description{
Generate a uniform 2D grid around space defined by x,y points
}
