% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_s3_uri}
\alias{is_s3_uri}
\title{Check if a path is an S3 URI}
\usage{
is_s3_uri(path)
}
\arguments{
\item{path}{A string representing a file path or S3 URI}
}
\value{
Boolean indicating whether the path is an S3 URI
}
\description{
Check if a path is an S3 URI
}
