% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{landscape_weights_2d}
\alias{landscape_weights_2d}
\title{Compute weight matrices to smooth gene expression signal along a 2D landscape}
\usage{
landscape_weights_2d(x, y, grid = grid_2d(x, y), k = 5, sd = 1)
}
\arguments{
\item{x}{x-axis position of the cells in the 2D space}

\item{y}{y-axis position of the cells in the 2D space}

\item{grid}{a 2-column tibble with x,y columns defining the grid where signal is estimated.}

\item{k}{number of nearest neighboring grid points to consider for each cell}

\item{sd}{standard deviation of the gaussian kernel}
}
\value{
a list of 2 matrices `fwd` and `rev`. That can be used to smooth gene expression signals.
        If m is a (gene x cells) matrix, the smoothed signal estimate at given timepoints is obtained
        with: m |> tcrossprod(w$fwd). The smoothed (gene x cells) expression matrix can be obtained
        with: m |> tcrossprod(w$fwd) |> tcrossprod(w$rev)
}
\description{
Weights are generated using a gaussian kernel on euclidian distances to
nearest neighbors.
}
\examples{
  x <- runif(1000); y <- runif(1000)
  m <- matrix(runif(1000*100),ncol=1000)
  w <- landscape_weights_2d(x,y,k=50,sd=0.05)
  M <- m |> tcrossprod(w$fwd) |> tcrossprod(w$rev)
  tibble(x,y,z=M[1,]) |> ggplot(aes(x=x,y=y,color=z)) + geom_point(size=3) + coord_equal()
}
