% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{lineage_ancestor_tbl}
\alias{lineage_ancestor_tbl}
\title{Compute ancestor table}
\usage{
lineage_ancestor_tbl(g)
}
\arguments{
\item{g}{a graph}
}
\value{
a sparse logical matrix where each row list all ancestors of a node

a 2-column tibble of ancestors
}
\description{
Compute ancestor table
}
