% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lineage_config}
\alias{lineage_config}
\title{Lineage Configuration}
\usage{
lineage_config(s3_bucket, region)
}
\arguments{
\item{s3_bucket}{S3 bucket name}

\item{region}{AWS region (e.g., "eu-west-1")}
}
\value{
A list containing the S3 bucket name, region, and AWS credentials
}
\description{
Lineage Configuration
}
