% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{lineage_coords}
\alias{lineage_coords}
\title{Compute cells coordinates in a lineage graph from their identity_matrix}
\usage{
lineage_coords(g, pseudotime, identity_scores)
}
\arguments{
\item{g}{a graph}

\item{identity_scores}{a numeric matrix of cell identity scores with number of
row matching length(pseudotime). Each row give identity scores of the cell
in each cluster.}
}
\value{
a tibble
}
\description{
Compute cells coordinates in a lineage graph from their identity_matrix
}
