% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_lineage.R
\name{lineage_graph_build}
\alias{lineage_graph_build}
\title{Build complete lineage graph with associated metrics to infer lineage}
\usage{
lineage_graph_build(pseudotime, identity_scores)
}
\arguments{
\item{pseudotime}{a numeric vector of pseudotime value for each cell}

\item{identity_scores}{a numeric matrix of cell identity scores with number of
row matching length(pseudotime). Each row give identity scores of the cell
in each cluster.}
}
\value{
a fully connected tbl_graph
}
\description{
Fit linear models between identity_scores and pseudotime, and compute
metrics to help identify edges with target identity increasing with time
and source indentity decreasing with time.
}
