% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{lineage_graph_prune}
\alias{lineage_graph_prune}
\title{Prune a lineage graph to keep one parent per node}
\usage{
lineage_graph_prune(g, n = 1L)
}
\arguments{
\item{g}{a lineage graph typically obtained with `lineage_graph_build`}

\item{n}{number of parent to keep for each node}
}
\value{
the filter input graph
}
\description{
Filter edges of the provided graph. First remove edges where median source node
pseudotime is later than median target node pseudotime. Then rank edges according
to lineage metrics and keep for each target node the best ranking parent.
}
