% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{main}
\alias{main}
\title{main}
\usage{
main(d = d, centers = 3, nmarker = 16)
}
\arguments{
\item{d}{A list containing 12 submatrices with different mutation types. Output of data_prepare(). Required.}

\item{centers}{Integer. The number of clusters used in Kmeans procedure. Default: 3.}

\item{nmarker}{Integer. The number of markers showed in final result. Default: 16.}
}
\value{
A list containing lineage analysis result.
}
\description{
main is the main part of LINEAGE for lineage analysis.
}
\details{
Main function
}
\examples{
data("TF1_clones")
data=TF1_clones$data
d=data_prepare(data)
a=dim(d[[1]])[2]
if(a>100){
    nmarker=c(15,20)
    centers=3
}else{
    nmarker=c(10,15)
    centers=2
}
result=main(d, centers, nmarker)

}
