% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multi_best_iter}
\alias{multi_best_iter}
\title{multi_best_iter}
\usage{
multi_best_iter(
  d,
  centers = c(2, 3),
  nmarker = c(10, 15, 20),
  repeats = 30,
  thread = 10
)
}
\arguments{
\item{d}{A list containing 12 submatrices with different mutation types. Output of data_prepare(). Required.}

\item{centers}{Integer. The number of clusters used in Kmeans procedure. Default: c(2,3).}

\item{nmarker}{Integer. The number of markers showed in final result. Default: c(10,15,20).}

\item{repeats}{Integer. The number of iterations. Default: 30.}

\item{thread}{Integer. Integer. The number of threads to run multi_best_iter. Default: 10.}
}
\value{
A list containing results of all repeats, including the best result.
}
\description{
multi_best_iter performs a parallel iteration process of main to guarantee a more stable and reliable cell clustering/clone tracing result.
}
\details{
Parallel iterative optimization
}
\examples{
data("TF1_clones")
data=TF1_clones$data
d=data_prepare(data)
a=dim(d[[1]])[2]
if(a>100){
    nmarker=c(15,20)
    centers=3
}else{
    nmarker=c(10,15)
    centers=2
}
# all_results=multi_best_iter(d, centers, nmarker, repeats=30, thread=10)

}
