% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{orm_balanced_cost_matrix}
\alias{orm_balanced_cost_matrix}
\title{Compute a balanced cost matrix with optional grouping}
\usage{
orm_balanced_cost_matrix(y, strata = NULL)
}
\arguments{
\item{y}{a factor of target labels}

\item{strata}{a factor of sample groups whose length match length(y). It is used
to update the cost matrix so that inter-strata penalties vanish.}
}
\description{
Grouping allow to train a single linear model performant on distinct set of samples like Male/Female
}
\examples{
orm_balanced_cost_matrix(iris$Species)
}
