% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{orm_fit}
\alias{orm_fit}
\title{Normalize an expression matrix and fit an ORM model}
\usage{
orm_fit(
  counts,
  y,
  C = NULL,
  normalize_fun = orm_lognorm_counts,
  strata = NULL,
  ...
)
}
\arguments{
\item{counts}{gene x sample expression matrix}

\item{y}{A factor defining the ordinal labels for each sample in x}

\item{C}{Cost matrix to penalize miss-predictions between labels, C[i,j]
being the cost for predicting label i instead of label j.
When NULL, it is initialized with a call to `orm_balanced_cost_matrix()`}

\item{normalize_fun}{Normalization function to use before fitting}

\item{strata}{when C is NULL, passed to `orm_balanced_cost_matrix()`}

\item{...}{additional parameters are passed to `orm_fit`}
}
\value{
linear model weights
}
\description{
Normalize an expression matrix and fit an ORM model
}
\examples{
library(tidyverse)
counts <- t(as.matrix(iris[1:4]))
y <- iris$Species \%>\% fct_relevel("versicolor","virginica","setosa")
w <- orm_fit(counts,y,LAMBDA=1e-3)
pred <- orm_predict(counts,w)
tibble(pred,y) \%>\% ggplot(aes(x=y,y=pred)) + geom_boxplot()
}
