% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{orm_lognorm_counts}
\alias{orm_lognorm_counts}
\title{Apply classical log-normalization on a raw counts}
\usage{
orm_lognorm_counts(counts, scaling_factor = 100)
}
\arguments{
\item{counts}{the expression matrix, generally given as raw counts}

\item{scaling_factor}{the normlized L1-norm of the samples}
}
\value{
log2(scale(counts,1/scaling_factor) + 1)
}
\description{
Apply classical log-normalization on a raw counts
}
\examples{
counts <- t(as.matrix(iris[1:4]))
logcounts <- orm_lognorm_counts(counts)
colSums(2^logcounts - 1)
}
