% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{orm_predict}
\alias{orm_predict}
\title{Normalize and return prediction of an ORM model for a given expression matrix}
\usage{
orm_predict(counts, w, normalize_fun = orm_lognorm_counts)
}
\arguments{
\item{counts}{gene x sample expression matrix with a maximum number of rownames
matching with names(w)}

\item{w}{a named vector of the model linear weight}

\item{normalize_fun}{Normalization function to use before fitting}
}
\value{
a numeric vector of predicted values for each column of counts
}
\description{
The expression matrix is first reduced to genes in the model, and normlized
}
\examples{
orm_balanced_cost_matrix(iris$Species)
}
