% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_s3_uri}
\alias{parse_s3_uri}
\title{Parse S3 URI into bucket and key}
\usage{
parse_s3_uri(s3_uri)
}
\arguments{
\item{s3_uri}{A string representing an S3 URI}
}
\value{
A list containing bucket and key
}
\description{
Parse S3 URI into bucket and key
}
