% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcluster.R
\docType{data}
\name{pcluster}
\alias{pcluster}
\title{pcluster example}
\format{
An object of class \code{"list"}
}
\source{
This dataset was generated during lineage analysis by LINEAGE.
}
\usage{
data("pcluster")
}
\description{
A list containing clustering results in selected subspaces with kmeans.
}
