% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{plot_lineage_graph}
\alias{plot_lineage_graph}
\title{Display lineage tree}
\usage{
plot_lineage_graph(g)
}
\arguments{
\item{g}{a lineage graph}
}
\value{
a gggraph plot
}
\description{
Display lineage tree
}
