% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_lineage.R
\name{plot_lineage_incidence_matrix}
\alias{plot_lineage_incidence_matrix}
\title{Show pseudotime/identity relationship}
\usage{
plot_lineage_incidence_matrix(g, pseudotime = NULL, identity_scores = NULL)
}
\arguments{
\item{g}{a lineage graph typically obtained with `lineage_graph_build`}

\item{pseudotime}{a numeric vector of pseudotime value for each cell}

\item{identity_scores}{a numeric matrix of cell identity scores with number of
row matching length(pseudotime). Each row give identity scores of the cell
in each cluster. The higher the score the more likely the cell belong to the cluster.}
}
\value{
a ggplot graph
}
\description{
Show pseudotime/identity relationship
}
