% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{register_directory}
\alias{register_directory}
\title{Register all files in a directory (input or output)}
\usage{
register_directory(session, dir_path, file_type, recursive = TRUE)
}
\arguments{
\item{session}{The lineage session object}

\item{dir_path}{Path to the directory (local) or S3 URI}

\item{file_type}{Either "inputs" or "outputs"}

\item{recursive}{Whether to recursively traverse subdirectories (default: TRUE)}
}
\value{
An updated session object
}
\description{
Register all files in a directory (input or output)
}
