% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{register_file}
\alias{register_file}
\title{Register a file (input or output)}
\usage{
register_file(session, file_path, file_type)
}
\arguments{
\item{session}{The lineage session object}

\item{file_path}{Path to the file (local or S3 URI)}

\item{file_type}{Either "inputs" or "outputs"}
}
\value{
An updated session object
}
\description{
Register a file (input or output)
}
