% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{register_input}
\alias{register_input}
\title{Register an input file}
\usage{
register_input(session, input_path)
}
\arguments{
\item{session}{The lineage session object}

\item{input_path}{Path to the input file}
}
\value{
An updated session object
}
\description{
Register an input file
}
