% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{register_input_directory}
\alias{register_input_directory}
\title{Register an input directory}
\usage{
register_input_directory(session, input_dir_path, recursive = TRUE)
}
\arguments{
\item{session}{The lineage session object}

\item{input_dir_path}{Path to the input directory}

\item{recursive}{Whether to recursively traverse subdirectories (default: TRUE)}
}
\value{
An updated session object
}
\description{
Register an input directory
}
