% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{register_output}
\alias{register_output}
\title{Register an output file}
\usage{
register_output(session, output_path)
}
\arguments{
\item{session}{The lineage session object}

\item{output_path}{Path to the output file}
}
\value{
An updated session object
}
\description{
Register an output file
}
