% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{register_output_directory}
\alias{register_output_directory}
\title{Register an output directory}
\usage{
register_output_directory(session, output_dir_path, recursive = TRUE)
}
\arguments{
\item{session}{The lineage session object}

\item{output_dir_path}{Path to the output directory}

\item{recursive}{Whether to recursively traverse subdirectories (default: TRUE)}
}
\value{
An updated session object
}
\description{
Register an output directory
}
