% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scc.R
\docType{data}
\name{scc}
\alias{scc}
\title{scc example}
\format{
An object of class \code{"numeric"}
}
\source{
This dataset was generated during lineage analysis by LINEAGE.
}
\usage{
data("scc")
}
\description{
The consensus clustering result.
}
