% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\docType{data}
\name{score}
\alias{score}
\title{score example}
\format{
An object of class \code{"numeric"}
}
\source{
This dataset was generated during lineage analysis by LINEAGE.
}
\usage{
data("score")
}
\description{
A dataframe containing adjusted rand indexes between the 
refined clustering results (resulted from t-SNE or UMAP-Kmeans procedures) 
and the clustering results in the selected subspaces.
}
