% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sum_score}
\alias{sum_score}
\title{sum_score}
\usage{
sum_score(score, cs, aris)
}
\arguments{
\item{score}{A dataframe containing adjusted rand indexes between the refined clustering results (resulted from t-SNE or UMAP-Kmeans procedures) and the clustering results in the selected subspaces. Required.}

\item{cs}{Indexes of selected subspaces. Required.}

\item{aris}{A dataframe containing ARI values among effective subspaces. Required.}
}
\value{
A list containing Sscore, suggested dimensional reduction method, number of effective subspaces and average ARI values used in Dscore calculation.
}
\description{
Calculate Sscore and Dscore for iterative optimization.
}
\details{
Sscore and Dscore calculation for optimization
}
\examples{
data("score")
data("cs")
data("aris")
score_sug=sum_score(score, cs, aris)

}
