% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF1_clones.R
\docType{data}
\name{TF1_clones}
\alias{TF1_clones}
\title{LINEAGE example and test dataset}
\format{
An object of class \code{"list"}
}
\source{
This dataset was generated from scRNA-seq data and annotation files obtained from NCBI.
}
\usage{
data("TF1_clones")
}
\description{
An example mitochondrial genotype matrix and its validated cell lineage information, which contain 70 cells with exogenous barcoding from 3 clones.
}
