% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{top_sites}
\alias{top_sites}
\title{top_sites}
\usage{
top_sites(d)
}
\arguments{
\item{d}{A dataframe containing mitochondrial variants frequency matrix where a column represented a single cell and a row represented variants frequency of a specific mitochondrial genotype. Required.}
}
\value{
A dataframe containing 2 modified mitochondrial genotype matrix with 20/50 highest variable sites.
}
\description{
Using mitochondrial variants frequency matrix as input, the 20/50 highest variable sites were called by identifying the rows with highest standard deviation (sd) across columns(cells).
}
\details{
Highly variable sites identification
}
\examples{
data("TF1_clones")
data=TF1_clones$data
d=data_prepare(data)
tops=data.frame()
tops2=data.frame()
for(i in 1:12){
    di=d[[i]]
    tops=rbind(tops,top_sites(di)[[1]])
    tops2=rbind(tops2,top_sites(di)[[2]])
}

}
