% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceplot.R
\name{traceplot}
\alias{traceplot}
\title{traceplot}
\usage{
traceplot(result, label)
}
\arguments{
\item{result}{The result of best_iter or multi_best_iter. Required.}

\item{label}{A dataframe containing lineage label with at least 2 columns named "Sample" and "Label". "Sample" column contains sample ids; "label" column is the lineage label of each sample. Required.}
}
\value{
A list containing a heatmap and a 2D visualization plot for the result of LINEAGE.
}
\description{
Using the result of best_iter or multi_best_iter as input, traceplot will export a heatmap and a suggested 2D visualization plot for the lineage analysis result.
}
\details{
Visualization of the result
}
\examples{
data("TF1_clones")
data=TF1_clones$data
rlabel=TF1_clones$rlabel
## performs a non-parallel iteration process 
result=lineage(data, repeats=30, thread=NULL)
## performs a parallel iteration process
# result=lineage(data, repeats=30, thread=10)
## plots with inferred clone labels
plots=traceplot(result = result, label= result$label)
## plots with reference clone labels
plots=traceplot(result = result, label= rlabel)

}
