% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{waves_weights_1d}
\alias{waves_weights_1d}
\title{Compute weight matrices to smooth gene expression signal along a pseudotime}
\usage{
waves_weights_1d(t, at = seq(min(t), max(t), length.out = 30), sd = 1)
}
\arguments{
\item{t}{cells pseudotime}

\item{at}{timepoints at which signal is evaluated}

\item{sd}{standard deviation of the gaussian kernel}
}
\value{
a list of 2 matrices `fwd` and `rev`. That can be used to smooth gene expression signals.
        If m is a (gene x cells) matrix, the smoothed signal estimate at given timepoints is obtained
        with: m |> tcrossprod(w$fwd). The smoothed (gene x cells) expression matrix can be obtained
        with: m |> tcrossprod(w$fwd) |> tcrossprod(w$rev)
}
\description{
Weights are generated using a gaussian kernel.
}
\examples{
t <- runif(100)
m <- matrix(runif(1000*100),ncol=100)
w <- waves_weights_1d(t)
m |> tcrossprod(w$fwd) |> tcrossprod(w$rev)
}
