% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_lineage.R
\name{lineage_graph_prune}
\alias{lineage_graph_prune}
\title{Prune a dense lineage graph to produce a lineage tree}
\usage{
lineage_graph_prune(g)
}
\arguments{
\item{g}{a lineage graph typically obtained with `lineage_graph_build`}
}
\value{
the filter input graph
}
\description{
Filter edges of the provided graph. First remove edges where median source node
pseudotime is later than median target node pseudotime. Then rank edges according
to lineage metrics and keep for each target node the best ranking parent.
}
