% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{orm_fit_pruned}
\alias{orm_fit_pruned}
\title{Two pass fitting of a ORM model with pruning}
\usage{
orm_fit_pruned(counts, ..., n = 25L)
}
\arguments{
\item{counts}{gene x sample expression matrix}

\item{...}{additional parameters are passed to `orm_fit`, and should at least include y}

\item{n}{Number of gene to keep after pruning (n positive + n negative)}
}
\value{
linear model weights
}
\description{
Fit a first ORM model with orm_fit() and select n top and n bottom features.
Then fit a second ORM model on selected features.
Note that in the second pass, the expression matrix is re-normlized only
considering selected features.
}
\examples{
library(tidyverse)
counts <- t(as.matrix(iris[1:4]))
y <- iris$Species \%>\% fct_relevel("versicolor","virginica","setosa")
w <- orm_fit_pruned(counts,y,n=1,LAMBDA=1e-3)
pred <- orm_predict(counts,w)
tibble(pred,y) \%>\% ggplot(aes(x=y,y=pred)) + geom_boxplot()
}
